remove dummy data, move TableData to types
